+++ packages/spacecat-shared-utils/src/index.js
@@ -12,26 +12,28 @@
 
 export {
   arrayEquals,
+  dateAfterDays,
+  deepEqual,
   hasText,
   isArray,
   isBoolean,
   isInteger,
-  isValidDate,
   isIsoDate,
   isIsoTimeOffsetsDate,
+  isNonEmptyArray,
+  isNonEmptyObject,
   isNumber,
   isObject,
-  isNonEmptyObject,
   isString,
-  toBoolean,
+  isValidDate,
   isValidUrl,
-  dateAfterDays,
-  deepEqual,
+  toBoolean,
 } from './functions.js';
 
 export {
   resolveSecretsName,
   resolveCustomerSecretsName,
+  getRUMDomainKey,
   generateCSVFile,
 } from './helpers.js';